size_t

+++ pdfgen.c
@@ -1965,7 +1965,7 @@ static int pdf_add_barcode_39(struct pdf_doc *pdf, struct pdf_object *page,
                               int x, int y, int width, int height,
                               const char *string, uint32_t colour)
 {
-    int len = strlen(string);
+    size_t len = strlen(string);
     int char_width = width / (len + 2);
 
     x = pdf_barcode_39_ch(pdf, page, x, y, char_width, height, colour, '*');
